Move testdata/ to internal/
